refactor (admin-verification-detail): separate detail page into some components
