+++ src/views/immunization/detail/LetterSection.vue
@@ -1,0 +1,64 @@
+<template>
+  <div class="d-flex flex-column">
+    <!-- Section Title -->
+    <div class="d-flex flex-row align-center mb-6 mt-8">
+      <span class="letter__section-title">{{ $t('label.applicant_letter') }}</span>
+      <img
+        src="/img/icons/arrow-down.svg"
+        alt="arrow-down"
+        height="18px"
+        @click="onClick"
+      >
+    </div>
+
+    <!-- Letter Field -->
+    <DisabledField
+      v-model="letter"
+      :label="$t('label.letter_number')"
+      class="mb-8"
+    >
+      <template #append>
+        <JDSButton class="ml-6" height="38px" @click="onDownload">
+          {{ $t('label.download') }}
+        </JDSButton>
+      </template>
+    </DisabledField>
+  </div>
+</template>
+
+<script>
+import JDSButton from '@/components/Base/JDSButton'
+import DisabledField from '@/components/Base/DisabledField'
+export default {
+  components: {
+    DisabledField,
+    JDSButton
+  },
+  props: {
+    letter: {
+      type: String,
+      default: ''
+    }
+  },
+  methods: {
+    onDownload() {
+      // @todo: create onDownload function
+    },
+    onClick() {
+      // @todo: create onClick function
+    }
+  }
+}
+</script>
+
+<style lang="scss" scoped>
+.letter {
+  &__section-title {
+    font-family: 'Roboto', sans-serif;
+    font-size: 24px;
+    font-weight: 700;
+    color: #BDBDBD;
+    margin-right: 15px;
+  }
+}
+</style>
